+++ Packages/Tracking/OpenXR/Runtime/Scripts/OpenXRLeapProvider.cs
@@ -53,13 +53,13 @@ namespace Ultraleap.Tracking.OpenXR
         {
             PopulateLeapFrame(ref _updateFrame);
 
-            Pose trackerTransform = new Pose(Vector3.zero, Quaternion.identity);
+            LeapTransform trackerTransform = new LeapTransform(Vector3.zero, Quaternion.identity, Vector3.one);
 
             // Adjust for relative transform if it's in use.
             var trackedPoseDriver = mainCamera.GetComponent<UnityEngine.SpatialTracking.TrackedPoseDriver>();
             if (trackedPoseDriver != null && trackedPoseDriver.UseRelativeTransform)
             {
-                trackerTransform.position += trackedPoseDriver.originPose.position;
+                trackerTransform.translation += trackedPoseDriver.originPose.position;
                 trackerTransform.rotation *= trackedPoseDriver.originPose.rotation;
             }
 
@@ -67,13 +67,16 @@ namespace Ultraleap.Tracking.OpenXR
             var parentTransform = mainCamera.transform.parent;
             if (parentTransform != null)
             {
-                trackerTransform.position += parentTransform.position;
+                trackerTransform.translation += parentTransform.position;
                 trackerTransform.rotation *= parentTransform.rotation;
+                trackerTransform.scale = parentTransform.lossyScale;
+            }
+            else
+            {
+                trackerTransform.scale = transform.lossyScale;
             }
 
-            _currentFrame = _updateFrame.TransformedCopy(new LeapTransform(
-                trackerTransform.position,
-                trackerTransform.rotation));
+            _currentFrame = _updateFrame.TransformedCopy(trackerTransform);
 
             DispatchUpdateFrameEvent(_currentFrame);
         }
@@ -119,6 +122,8 @@ namespace Ultraleap.Tracking.OpenXR
                 return false;
             }
 
+
+            float timeVisible = 0;
             if (handTracker == HandTracker.Left)
             {
                 if (_leftHandFirstSeen_ticks == -1)
@@ -126,6 +131,7 @@ namespace Ultraleap.Tracking.OpenXR
                     _leftHandFirstSeen_ticks = DateTime.Now.Ticks;
                     _leftHandId = _handId++;
                 }
+                timeVisible = ((float)(DateTime.Now.Ticks - _leftHandFirstSeen_ticks)) / (float)TimeSpan.TicksPerSecond;
             }
             else
             {
@@ -134,6 +140,8 @@ namespace Ultraleap.Tracking.OpenXR
                     _rightHandFirstSeen_ticks = DateTime.Now.Ticks;
                     _rightHandId = _handId++;
                 }
+                timeVisible = ((float)(DateTime.Now.Ticks - _rightHandFirstSeen_ticks)) /
+                              (float)TimeSpan.TicksPerSecond;
             }
 
             for (int fingerIndex = 0; fingerIndex < 5; fingerIndex++)
@@ -186,7 +194,7 @@ namespace Ultraleap.Tracking.OpenXR
                     _frameId,
                     (handTracker == HandTracker.Left ? 0 : 1),
                     fingerIndex,
-                    10f, // Fixed for now
+                    timeVisible,
                     joints[xrTipIndex].Pose.position,
                     (joints[xrTipIndex].Pose.rotation * Vector3.forward),
                     fingerWidth,
@@ -207,8 +215,7 @@ namespace Ultraleap.Tracking.OpenXR
                 CalculatePinchDistance(ref hand),
                 palmWidth,
                 handTracker == HandTracker.Left,
-                handTracker == HandTracker.Left ? ((float)(DateTime.Now.Ticks - _leftHandFirstSeen_ticks)) / (float)TimeSpan.TicksPerSecond :
-                                                  ((float)(DateTime.Now.Ticks - _rightHandFirstSeen_ticks)) / (float)TimeSpan.TicksPerSecond,
+                timeVisible,
                 null, // Already Populated
                 joints[(int)HandJoint.Palm].Pose.position,
                 joints[(int)HandJoint.Palm].Pose.position,