+++ Packages/Tracking/CHANGELOG.md
@@ -16,8 +16,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Detector null reference error when creating a detector at runtime
-- XRServiceProvider does not scale when the player scales
 - InteractionSlider now raises event for value changes when setting values via the Horizontal and Vertical Percent properties
+- XRServiceProvider and OpenXRLeapProvider do not scale when the player scales
+- `timeVisible` was not populated on the OpenXR Provider for `Finger`s
 
 ### Known issues 
 - Offset between skeleton hand wrist and forearm in sample scenes